added not foun validation

+++ app/Http/Controllers/Country/CountryController.php
@@ -14,7 +14,11 @@ class CountryController extends Controller
     }
 
     public function findCountryById($id){
-        return response()->json(CountryModel::find($id) , 200);
+        $country = CountryModel::find($id);
+        if (is_null($country)){
+            return response()->json("Not found data" , 404);
+        }
+        return response()->json($country, 200);
     }
 
     public function addCountry(Request $request){
@@ -24,12 +28,18 @@ class CountryController extends Controller
 
     public function updateCountry($id,Request $request ){
 //        $country->update($request->all($id));
+        if (is_null(CountryModel::find($id))){
+            return response()->json("Not found data" , 404);
+        }
         CountryModel::where('id' ,$id)->update($request->all());
         $country=CountryModel::find($id);
         return response()->json($country ,200);
     }
 
     public function deleteCountryById($id,Request $request ){
+        if (is_null(CountryModel::find($id))){
+            return response()->json("Not found data" , 404);
+        }
         CountryModel::where('id' ,$id)->delete();
         return response()->json('deleted' , 200);
     }
